Rename project to Mapper and clean up
