+++ Mapper.sln
@@ -3,8 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
- main
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "InputToControllerMapper", "InputToControllerMapper\InputToControllerMapper.csproj", "{4E4BAA4A-FE7E-4F5A-80EB-4E6E90722663}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Mapper", "Mapper\Mapper.csproj", "{4E4BAA4A-FE7E-4F5A-80EB-4E6E90722663}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,7 +14,6 @@ Global
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-main
 		{4E4BAA4A-FE7E-4F5A-80EB-4E6E90722663}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{4E4BAA4A-FE7E-4F5A-80EB-4E6E90722663}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4E4BAA4A-FE7E-4F5A-80EB-4E6E90722663}.Release|Any CPU.ActiveCfg = Release|Any CPU
