+++ Mapper/ApplicationConfiguration.cs
@@ -1,6 +1,6 @@
 using System.Windows.Forms;
 
-namespace InputToControllerMapper;
+namespace Mapper;
 
 internal static class ApplicationConfiguration
 {
