+++ README.md
@@ -8,11 +8,10 @@ A .NET 9 solution to log mouse and Wooting analog keyboard input using a WinForm
 - Wooting Analog SDK: see `SDK-PLUGINS.txt` for download links and instructions.
 
 ## Solution Structure
-- `InputMappingSolution.sln` – solution file.
-main
+- `Mapper.sln` – solution file.
   - `ApplicationConfiguration.cs`
   - `Program.cs`
-  - `InputToControllerMapper.csproj` (copies DLLs).
+  - `Mapper.csproj` (copies DLLs).
 - `native_sdks/` – drop all native DLLs here (initially empty).
 - `SDK-PLUGINS.txt` – SDK setup and troubleshooting guide.
 - `README.md` – this file.
@@ -21,8 +20,7 @@ main
 1. Install the .NET 9 SDK.
 2. Download the Wooting Analog SDK and place `wooting_analog_wrapper.dll` in `native_sdks`.
 3. Build the solution with `dotnet build` or open it in Visual Studio.
-4. Run the `InputToControllerMapper` project and watch the textbox for mouse and analog keyboard logs.
- main
+4. Run the `Mapper` project and watch the textbox for mouse and analog keyboard logs.
 
 ## Troubleshooting
 If input events or the SDK fail to load, review `SDK-PLUGINS.txt` for hints about DLL placement and architecture.
